Create stm32f407xx.h
This header file contains macros for addresses of peripherals, registers and other general macros that will be used in developing driver files. Please note that this file also contains macros for peripherals other than the GPIO driver too because this is a generic file. It will be used for the development of all drivers