Create stm32f407xx_gpio_driver.c
This is the driver file for GPIO driver. It includes all the necessary functions required to configure the GPIO driver. Please remember this is intended only for STM32F4xx based MCUs. It makes use of the macros defined in the gpio driver header file  as well as the MCU specific header file.