Create LED Toggle with External Btn.c
This code file achieves the on-board LED toggling using an external button rather than the on-board button. The external button is connected on the PB12 pin. The pulled up resistor is activated so the button press pulls the pin own. This is probed by the if block at the end of the code.